+++ src/components/index.ts
@@ -4,4 +4,4 @@ export { Skills, Summary } from './about-me';
 export { default as SubTitle } from './SubTitle';
 export { default as SubTitleEdit } from './SubTitleEdit';
 export { GithubImg, GoogleImg, NaverImg } from './SnsIcon';
-export { LoginButton, SnsLoginButton } from './login';
+export { LoginButton, SnsLoginButton, GoogleLoginButton, GoogleSignUpIcon } from './login';
