+++ src/components/login/index.tsx
@@ -4,3 +4,4 @@ import 정리용 페이지
 
 export { LoginButton } from './LoginButton';
 export { SnsLoginButton } from './SnsLoginButton';
+export { GoogleLoginButton, GoogleSignUpIcon } from './GoogleButton';
