+++ src/pages/login/Login.tsx
@@ -1,11 +1,10 @@
 import styled from 'styled-components';
 import { LoginContainer } from './LoginContainer';
-import { GithubImg, GoogleImg, NaverImg, SnsLoginButton } from '../../components';
+import { GithubImg, NaverImg, SnsLoginButton, GoogleLoginButton, GoogleSignUpIcon } from '../../components';
 import { useSetRecoilState } from 'recoil';
 import { pageState } from '../../utils/data/atom';
 import { useEffect } from 'react';
 import { LoginSpace } from 'LoginModule';
-import { GoogleLogin } from 'react-google-login';
 
 const Login = () => {
 
@@ -23,24 +22,13 @@ const Login = () => {
     const GITHUB_CLIENT_ID = process.env.REACT_APP_GITHUB_CLIENT_ID || "";
     const githubUri: string = `https://github.com/login/oauth/authorize?client_id=${GITHUB_CLIENT_ID}&redirect_uri=${oauthRedirect}&scope=user:email%20read:user&state=1`
 
-    // google
-    const GOOGLE_CLIENT_ID = process.env.REACT_APP_GOOGLE_CLIENT_ID || "";
-    const googleUri: string = `https://accounts.google.com/o/oauth2/v2/auth?response_type=code&client_id=${GOOGLE_CLIENT_ID}&redirect_uri=${oauthRedirect}&scope=openid%20profile%20email&access_type=offline&state=1`
-
     function handleClick(event: any) {
         if (event.target.name === "naverIcon") {
             window.open(naverUri, "_self");
         } else if (event.target.name === "githubIcon") {
             window.open(githubUri, "_self");
-        } else {
-            window.open(googleUri, "_self");
         }
     }
-
-    const responseGoogle = (response: any) => {
-        console.log('response', response);
-        console.log('profile', response.getAuthResponse());
-      }
 
     // 회원가입 페이지 번호 리셋
     useEffect(() => setPage(0));
@@ -49,25 +37,12 @@ const Login = () => {
         <LoginContainer>
             <Logo>EliceFolio</Logo>
             <SnsLoginButton text='깃허브로 로그인' to="github" color="black" />
-            <GoogleLogin
-                clientId={GOOGLE_CLIENT_ID}
-                onSuccess={responseGoogle}
-                onFailure={responseGoogle}
-                cookiePolicy='single_host_origin'
-                render={ renderProps => (
-                    <GoogleButton
-                        onClick={renderProps.onClick}
-                        disabled={renderProps.disabled}>
-                        <GoogleImg />
-                        구글로 로그인
-                    </GoogleButton>
-                )}
-            />
+            <GoogleLoginButton />
             <SnsLoginButton text='네이버로 로그인' to='naver' color="#19CE60" />
             <TextP>회원이 아니신가요?</TextP>
             <IconDiv onClick={(event: any) => handleClick(event)}>
                 <GithubImg size="20%" />
-                <GoogleImg size="40%" />
+                <GoogleSignUpIcon />
                 <NaverImg size="20%" />
             </IconDiv>
         </LoginContainer>
@@ -89,38 +64,6 @@ const Logo = styled.p`
     @media screen and (max-height: 340px) {
     margin-bottom: 1vh;
 }
-`;
-
-// 구글 버튼
-const GoogleButton = styled.button`
-    all: unset;
-
-    position: relative;
-
-    width: 19.5vw;
-    height: 6.7vh;
-
-    margin: 5px 0;
-
-    border-radius: 10px;
-
-    text-align: center;
-    font-family: 'AppleSDGothicNeo', 'sans-serif';
-    color: white;
-    font-size: 1vw;
-
-    background-color: #EA4335;
-
-    cursor: pointer;
-
-    & img {
-        position: absolute;
-
-        top: 10px;
-        left: 30px;
-
-        filter: invert(96%) sepia(99%) saturate(3%) hue-rotate(335deg) brightness(104%) contrast(100%);
-    }
 `;
 
 // 아이콘 영역
